+++ app/routers/resources.py
@@ -1,7 +1,7 @@
 from typing import Optional, List
 from fastapi import FastAPI, Response, status, HTTPException, APIRouter, Depends
 from sqlalchemy.orm import Session
-from .. import schemas, models, database, utils
+from .. import schemas, models, database, utils, oauth2
 
 router = APIRouter(
     prefix="/resources",
@@ -17,7 +17,10 @@ def get_resources(db: Session = Depends(database.get_db)):
     return resources
 
 @router.post("/", response_model=schemas.ResourceOut)
-def create_resource(resource: schemas.ResourceBase, status_code=status.HTTP_201_CREATED, db: Session = Depends(database.get_db)):
+def create_resource(resource: schemas.ResourceBase, status_code=status.HTTP_201_CREATED, db: Session = Depends(database.get_db), current_user: int = Depends(oauth2.get_current_user)):
+
+    if not current_user.admin:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail=f"User with id: {current_user.id} is not an admin")
 
     new_resource = models.Resource(**resource.dict())
     db.add(new_resource)
@@ -37,13 +40,16 @@ def get_resource(id: int, db: Session = Depends(database.get_db)):
     return resource
 
 @router.put("/{id}", response_model=schemas.ResourceOut)
-def update_resource(id: int, updated_resource: schemas.ResourceBase, db: Session = Depends(database.get_db)):
+def update_resource(id: int, updated_resource: schemas.ResourceBase, db: Session = Depends(database.get_db), current_user: int = Depends(oauth2.get_current_user)):
 
     resource_query = db.query(models.Resource).filter(models.Resource.id == id)
     resource = resource_query.first()
 
     if not resource:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Resource with id: {id} does not exist")
+
+    if not current_user.admin:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail=f"User with id: {current_user.id} is not an admin")
 
     resource_query.update(updated_resource.dict(), synchronize_session=False)
     db.commit()
@@ -51,12 +57,15 @@ def update_resource(id: int, updated_resource: schemas.ResourceBase, db: Session
     return resource
 
 @router.delete("/{id}")
-def delete_resources(id: int, db: Session = Depends(database.get_db)):
+def delete_resources(id: int, db: Session = Depends(database.get_db), current_user: int = Depends(oauth2.get_current_user)):
     resource_query = db.query(models.Resource).filter(models.Resource.id == id)
     resource = resource_query.first()
 
     if not resource:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Resource with id: {id} does not exist")
+
+    if not current_user.admin:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail=f"User with id: {current_user.id} is not an admin")
 
     resource_query.delete(synchronize_session=False)
     db.commit()
