+++ app/routers/users.py
@@ -1,7 +1,7 @@
 from typing import Optional, List
 from fastapi import FastAPI, Response, status, HTTPException, APIRouter, Depends
 from sqlalchemy.orm import Session
-from .. import schemas, models, database, utils
+from .. import schemas, models, database, utils, oauth2
 
 router = APIRouter(
     prefix="/users",
@@ -10,9 +10,12 @@ router = APIRouter(
 
 #GET USERS - gets all the users 
 @router.get("/", response_model=List[schemas.UserOut])
-def get_users(db: Session = Depends(database.get_db)):
+def get_users(db: Session = Depends(database.get_db), current_user: int = Depends(oauth2.get_current_user)):
 
     users = db.query(models.User).all()
+
+    if not current_user.admin:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail=f"User with id: {current_user.id} is not an admin")
 
     return users
 
@@ -38,25 +41,31 @@ def create_user(user: schemas.UserBase, status_code=status.HTTP_201_CREATED, db:
 
 #GET USER - gets a user with a certain id
 @router.get("/{id}", response_model=schemas.UserOut)
-def get_user(id: int, db: Session = Depends(database.get_db)):
+def get_user(id: int, db: Session = Depends(database.get_db), current_user: int = Depends(oauth2.get_current_user)):
 
     user = db.query(models.User).filter(models.User.id == id).first() 
 
     if not user:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id: {id} was not found")
 
+    if id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail=f"Unauthorized to access user with id: {id}")  
+
     return user
 
 
 #DELETE USER - deletes a user with certain id
 @router.delete("/{id}")
-def delete_user(id: int, db: Session = Depends(database.get_db)):
+def delete_user(id: int, db: Session = Depends(database.get_db), current_user: int = Depends(oauth2.get_current_user)):
 
     user_query = db.query(models.User).filter(models.User.id == id)
     user = user_query.first()
 
     if user == None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id: {id} was not found") 
+
+    if id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail=f"Unauthorized to access user with id: {id}")  
 
     user_query.delete(synchronize_session=False)
     db.commit()
@@ -66,7 +75,7 @@ def delete_user(id: int, db: Session = Depends(database.get_db)):
 
 #UPDATE USER - updates the user
 @router.put("/{id}", response_model=schemas.UserOut)
-def update_user(id: int, updated_user: schemas.UserBase, db: Session = Depends(database.get_db)):
+def update_user(id: int, updated_user: schemas.UserUpdate, db: Session = Depends(database.get_db), current_user: int = Depends(oauth2.get_current_user)):
 
     user_query = db.query(models.User).filter(models.User.id == id)
     user = user_query.first()
@@ -74,6 +83,10 @@ def update_user(id: int, updated_user: schemas.UserBase, db: Session = Depends(d
     if user == None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"User with id: {id} was not found")
 
+    if id != current_user.id:
+       raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail=f"Unauthorized to access user with id: {id}") 
+
     user_query.update(updated_user.dict(), synchronize_session=False)
+    db.commit()
 
     return user
